fix(gfx): widen asset atlas icon_uid to uint16_t; fix sentinel overflow
- tools/asset_atlas.py: emit uint16_t icon_uid in AssetAtlasEntry and
  asset_atlas_icon_tile accessor to support 501 unique icon tiles
  (uids 0..500) without truncation. Use pal = 255 (0xFF) sentinel for
  uint8_t icon_pal field in emit_entries_inc.
- src/gfx/asset_atlas.h: regenerate header with uint16_t icon_uid.
- src/gfx/asset_atlas.c: update asset_atlas_icon_tile signature to uint16_t uid.
- src/gfx/asset_atlas_entries.inc: regenerate entries with correct sentinels.

Verified: make atlas, make atlas-check, make lint (0 warnings), make memmap,
make test, make test-harness (172/172 PASS), make verify-oam all pass.

+++ src/gfx/asset_atlas.c
@@ -26,7 +26,7 @@ void asset_atlas_get(AssetId id, AssetAtlasEntry *out)
     *out = s_atlas_scratch;
 }
 
-void asset_atlas_icon_tile(uint8_t uid, uint8_t *out)
+void asset_atlas_icon_tile(uint16_t uid, uint8_t *out)
 {
     banked_copy(ASSET_ATLAS_BANK_ICONS, out,
                 &g_asset_icon_tiles[(uint16_t)uid << 4], 16);
